pyproject.toml
[tool.poetry]
name = "rainbowlog"
version = "1.0.0"
description = "Format your python logs with colours based on the log levels."
authors = ["Abraham Murciano <abrahammurciano@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/abrahammurciano/rainbowlog"
documentation = "https://abrahammurciano.github.io/rainbowlog/rainbowlog/"
keywords = ["logging", "color"]

[tool.poetry.dependencies]
python = "^3.7"
ansicolors = "^1.1"
importlib-metadata = "^4.11"

[tool.poetry.dev-dependencies]
pytest = ">=7.1"
pdoc3 = ">=0.10.0"
types-toml = ">=0.10.7"
toml = ">=0.10.2"
black = {version = ">=22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
